move files and fix routing
