+++ frontend/src/components/Nav.tsx
@@ -1,16 +1,16 @@
 import React from "react";
-import { Link } from "react-router-dom";
+import NavItem from "./NavItem";
 
 const Nav: React.FC = () => {
   return (
     <nav>
       <ul>
-        <li><Link to="/home">Home</Link></li>
-        <li><Link to="/newdr">New Dr</Link></li>
-        <li><Link to="/newrx">New Rx</Link></li>
-        <li><Link to="/rxitem">Rx Item</Link></li>
-        <li><Link to="/sigcode">Sig Code</Link></li>
-        <li><Link to="/x">X</Link></li>
+        <NavItem name="Home" path="/" />
+        <NavItem name="NewDr" path="/NewDr" />
+        <NavItem name="NewRx" path="/NewRx" />
+        <NavItem name="RxItem" path="/RxItem" />
+        <NavItem name="SigCode" path="/SigCode" />
+        <NavItem name="X" path="/X" />
       </ul>
     </nav>
   );
