+++ frontend/src/pages/RxItem.tsx
